speed send is now cleaned up
